Add files via upload
This is the frontend of PULSE, a mental health and wellbeing app designed to support emotional stability and boost productivity. Built with HTML, CSS, and JavaScript, it delivers a clean, responsive, and user-friendly interface. The design focuses on simplicity, accessibility, and engaging visuals to encourage positive user interaction. This frontend is structured to integrate smoothly with backend services, enabling features like mood tracking, productivity tools, and personalized wellbeing resources.